breaking up webhook manager duties
